refactor: move prover
